522: v1.15: add capacity route
